fix(config): address CodeRabbit round 2 findings
Replace hardcoded "256" literal in riic_peripheral_write log message with
symbolic reference to k_riic_peripheral_transfer_limit. Add missing wire_len
size guard in internal_handle_reset before riic_peripheral_write (was only
present in internal_send_frame). Remove redundant (uint32_t) casts when
assigning k_encoded_len_none in nanopb encode stubs. Add full Doxygen
documentation to the six previously undocumented uart_debug_* and
rx_log_usb_putc/puts stubs in mock_rx_log.c.

+++ star-rx72n-firmware/tests/mocks/mock_comm_transports.c
@@ -419,7 +419,7 @@ rx_err_t rx_nanopb_encode_velocity_response(const void* msg,
   (void)buffer;
   (void)buffer_size;
   if (len != nullptr) {
-    *len = (uint32_t)k_encoded_len_none;
+    *len = k_encoded_len_none;
   }
   return k_rx_ok;
 }
@@ -453,7 +453,7 @@ rx_nanopb_encode_estop_response(const void* msg, void* buffer, uint32_t buffer_s
   (void)buffer;
   (void)buffer_size;
   if (len != nullptr) {
-    *len = (uint32_t)k_encoded_len_none;
+    *len = k_encoded_len_none;
   }
   return k_rx_ok;
 }
@@ -489,7 +489,7 @@ rx_err_t rx_nanopb_encode_pid_gains_response(const void* msg,
   (void)buffer;
   (void)buffer_size;
   if (len != nullptr) {
-    *len = (uint32_t)k_encoded_len_none;
+    *len = k_encoded_len_none;
   }
   return k_rx_ok;
 }
